backend/requirements.txt
azure-common==1.1.25
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-nspkg==3.0.2
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
geographiclib==1.50
geopy==1.22.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
wincertstore==0.2
